Refactor code to beautiful
